Separate CertificationImageService, fix PhotoSelectionViewModel UIKit exposure, remove View coordination
closes #1, closes #2, closes #3

Also renames PhotoSelectionViewMdoel.swift → PhotoSelectionViewModel.swift (typo fix).
