+++ TimerStamp/Presentations/Screens/CertificationModalView.swift
@@ -15,13 +15,13 @@ struct CertificationModalView: View {
     @State private var composedImage: UIImage?
     @State private var showShareSheet = false
     @State private var showSaveConfirmation = false
-    
+
     var body: some View {
         VStack(spacing: 24) {
             Text(L10n.certTitle)
                 .font(.title2)
                 .foregroundStyle(.mainText)
-            
+
             if let image = composedImage {
                 Image(uiImage: image)
                     .resizable()
@@ -32,7 +32,7 @@ struct CertificationModalView: View {
                 ProgressView(L10n.certRendering)
                     .frame(height: 300)
             }
-            
+
             HStack(spacing: 16) {
                 Button(action: onDismiss) {
                     Text(L10n.certClose)
@@ -42,7 +42,7 @@ struct CertificationModalView: View {
                         .background(Color.mainButton)
                         .cornerRadius(10)
                 }
-                
+
                 Button(action: saveImage) {
                     Text(L10n.certSaveImage)
                         .foregroundColor(.mainButtonText)
@@ -52,7 +52,7 @@ struct CertificationModalView: View {
                         .cornerRadius(10)
                 }
                 .disabled(composedImage == nil)
-                
+
                 Button(action: {
                     showShareSheet = true
                 }) {
@@ -67,7 +67,7 @@ struct CertificationModalView: View {
             }
             .frame(height: 44)
             .padding(.horizontal)
-            
+
             Spacer()
         }
         .padding()
@@ -77,7 +77,7 @@ struct CertificationModalView: View {
                     if !hasRendered {
                         hasRendered = true
                         DispatchQueue.main.async {
-                            if let image = ComposedImageRenderer.render(image: baseImage, minutes: minutes) {
+                            if let image = CertificationImageService.render(image: baseImage, minutes: minutes) {
                                 DispatchQueue.main.async {
                                     composedImage = image
                                 }
@@ -92,7 +92,7 @@ struct CertificationModalView: View {
             }
         }
         .onReceive(NotificationCenter.default.publisher(for: .imageSaveCompleted)) { notification in
-            if let success = notification.object as? Bool {
+            if notification.object is Bool {
                 showSaveConfirmation = true
             }
         }
@@ -103,107 +103,15 @@ struct CertificationModalView: View {
         }
         .background(Color.mainBackground)
     }
-    
-    func saveImage() {
+
+    private func saveImage() {
         guard let image = composedImage else { return }
-        UIImageWriteToSavedPhotosAlbum(image, ImageSaveHelper.shared, #selector(ImageSaveHelper.didFinishSaving(_:didFinishSavingWithError:contextInfo:)), nil)
-    }
-    
-    var formattedDate: String {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .medium
-        formatter.timeStyle = .short
-        return formatter.string(from: Date())
-    }
-}
-
-class ImageSaveHelper: NSObject {
-    static let shared = ImageSaveHelper()
-    var onSaveCompletion: ((Bool) -> Void)?
-    
-    @objc func didFinishSaving(_ image: UIImage, didFinishSavingWithError error: Error?, contextInfo: UnsafeRawPointer) {
-        DispatchQueue.main.async {
-            let success = (error == nil)
-            NotificationCenter.default.post(name: .imageSaveCompleted, object: success)
-        }
-    }
-}
-
-extension Notification.Name {
-    static let imageSaveCompleted = Notification.Name("imageSaveCompleted")
-}
-
-enum ComposedImageRenderer {
-    @MainActor
-    static func render(image: UIImage, minutes: Int) -> UIImage? {
-        let view = ZStack {
-            GeometryReader { geo in
-                let timerSize = 400.0
-                let radius = timerSize / 3.4
-                let margin: CGFloat = 60
-                
-                // 1. 원본 이미지
-                Image(uiImage: image)
-                    .resizable()
-                    .scaledToFill()
-                    .frame(width: geo.size.width, height: geo.size.height)
-                
-                // 2. 그라데이션 오버레이
-                LinearGradient(
-                    gradient: Gradient(colors: [
-                        Color.black.opacity(0),
-                        Color.black.opacity(0.3)
-                    ]),
-                    startPoint: UnitPoint(x: 0.5, y: 0.7),
-                    endPoint: .bottom
-                )
-                .frame(width: geo.size.width, height: geo.size.height)
-                
-                // 3. 절대 위치로 하단 정렬
-                ZStack(alignment: .bottom) {
-                    Color.clear
-                        .frame(width: geo.size.width, height: geo.size.height)
-                    
-                    HStack(alignment: .bottom, spacing: 0) {
-                        // 좌측: 타이머
-                        VStack(alignment: .leading) {
-                            SimpleTimerView(minutes: minutes, radius: radius)
-                                .opacity(0.74)
-                                .frame(maxHeight: .infinity, alignment: .bottom)
-                                .padding(66)
-                        }
-                        
-                        Spacer(minLength: 20)
-           
-                        VStack(alignment: .trailing) {
-                            Text("✨ \(L10n.focusComplete(minutes))")
-                                .font(.system(size: 66, weight: .heavy))
-                                .fontWidth(.expanded)
-                                .kerning(-2)
-                                .foregroundColor(.white)
-                                .multilineTextAlignment(.trailing)
-                                .lineLimit(nil)
-                                .padding(.bottom, 6)
-                            
-                            Text(DateFormatter.localizedString(from: Date(), dateStyle: .medium, timeStyle: .short))
-                                .font(.system(size: 58, weight: .light))
-                                .foregroundColor(.white.opacity(0.95))
-                                .padding(.bottom, 10)
-                        }
-                        .frame(maxHeight: .infinity, alignment: .bottom)
-                    }
-                    .padding(.horizontal, margin)
-                    .padding(.bottom, margin)
-                }
-            }
-            .frame(width: 1080, height: 1920)
-        }
-        .background(Color.clear)
-        
-        let renderer = ImageRenderer(content: view)
-        renderer.proposedSize = .init(CGSize(width: 1080, height: 1920))
-        renderer.isOpaque = true
-        return renderer.uiImage
+        UIImageWriteToSavedPhotosAlbum(
+            image,
+            ImageSaveService.shared,
+            #selector(ImageSaveService.didFinishSaving(_:didFinishSavingWithError:contextInfo:)),
+            nil
+        )
     }
 }
 
@@ -216,21 +124,5 @@ enum ComposedImageRenderer {
         )
         .environment(\.locale, .init(identifier: "en"))
         .previewDisplayName("English")
-        
-//        CertificationModalView(
-//            baseImage: UIImage(named: "s2") ?? UIImage(),
-//            minutes: 45,
-//            onDismiss: {}
-//        )
-//        .environment(\.locale, .init(identifier: "ko"))
-//        .previewDisplayName("한국어")
-//        
-//        CertificationModalView(
-//            baseImage: UIImage(named: "s2") ?? UIImage(),
-//            minutes: 45,
-//            onDismiss: {}
-//        )
-//        .environment(\.locale, .init(identifier: "ja"))
-//        .previewDisplayName("日本語")
     }
 }
